Improve Playground.app
This now creates a skeleton folder for the playground and opens that
skeleton up with Xcode, rather than automating the UI process which can
be tedious (having to remember shortcuts) and slow (having to wait for
at least 10 seconds for the UI automation to complete).

With this commit, a new playground file is created and can be edited in
around 2 seconds, which is a drastic improvement.